requirements.txt
python-telegram-bot==20.7
python-binance==1.0.19
pandas==2.1.4
numpy==1.24.3
flask==3.0.0
requests==2.31.0